Super Light Weight Inventory Management System (SLIMS)
An Inventory Management System targeted at the Micro - Small Business
Segment.
These are the companies that do not need & can not afford full fledged
Inventory Management systems that are available in the market and hence
have to make use of manual techniques of book-keeping which are
time-consuming as well as error-prone.

This software makes use of the web browser therefore it is:
1. Simple to interpret
2. Has a flat learning curve
3. Point and click (Minimal keyboard usage)

The key features of this software are :
1. Keeps track of the current stock of each item
2. Notifies the manager if the stock of a certain item has dropped below
a certain threshold value.
3. Displays transactions according to Product, Customer, Supplier or
Date range (which ever option is chosen by the user.)